requirements.txt
Django==2.2.4
django-storages==1.7.1
gunicorn==19.9.0
psycopg2-binary==2.8.3
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
